+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#146](https://github.com/os2display/display-templates/pull/146)
+  - Fixed issue where image-text image cycle is not restarting when only on slide is displayed.
+  - Fixed issue with slideshow default duration being 5000 s.
+
 ## [1.3.1] - 2023-10-26
 
 - [#145](https://github.com/os2display/display-templates/pull/145)